+++ src/cli/cli.js
@@ -1,6 +1,7 @@
 import install from './commands/install.js'
 import add from './commands/add.js'
 import remove from './commands/remove.js'
+import clean from './commands/clean.js'
 
 import debug from './options/debug.js'
 import use from './options/use.js'
@@ -11,7 +12,8 @@ import info from './options/info.js'
 const commands = [
   install,
   add,
-  remove
+  remove,
+  clean
 ]
 
 const options = {